nu_table: Rename wrap.rs to width_control.rs
