+++ crates/nu-table/src/width_control.rs
